Revert "config: use "long int" as required by my version of libconfig"
This reverts commit 07bcc14e93564533261dbaec9080a383334323dc.

+++ wmediumd/config.c
@@ -179,7 +179,7 @@ int load_config(const char *file)
 	config_t cfg, *cf;
 	const config_setting_t *ids, *prob_list, *mat_array, *jammer_s;
 	int count_ids, rates_prob, i, j;
-	long int count_value, rates_value;
+	int count_value, rates_value;
 
 	/*initialize the config file*/
 	cf = &cfg;
